add topbar with commit/discard buttons (#627)
* add topbar with commit/discard buttons

* fix cache invalidation after commit/discard
