+++ packages/frinx-gamma/src/components/commit-status-modal/commit-status-modal.helpers.ts
@@ -1,4 +1,4 @@
-import { useEffect, useRef, useState } from 'react';
+import { useEffect, useState } from 'react';
 import uniflowCallbackUtils from '../../uniflow-callback-utils';
 
 export type TaskStatus = 'COMPLETED' | 'FAILED' | 'SCHEDULED' | 'IN_PROGRESS';
@@ -60,11 +60,14 @@ export type ExecutedWorkflowResponse<T> = {
   result: ExecutedWorkflowPayload<T>;
 };
 
-async function getWorkflowExecOutput<T>(workflowId: string, abortController: AbortController) {
-  const callbacks = uniflowCallbackUtils.getCallbacks;
-  const response = await callbacks.getWorkflowInstanceDetail(workflowId, { signal: abortController.signal });
-  const data = response as T;
-  return data;
+async function getWorkflowExecOutput<T>(workflowId: string, abortController: AbortController): Promise<T | null> {
+  const { getWorkflowInstanceDetail } = uniflowCallbackUtils.getCallbacks;
+  try {
+    const response = await getWorkflowInstanceDetail(workflowId, { signal: abortController.signal });
+    return response as T;
+  } catch {
+    return null;
+  }
 }
 
 export type AsyncGeneratorParams = {
@@ -79,13 +82,17 @@ export async function* asyncGenerator<T>({
   onFinish,
 }: AsyncGeneratorParams): AsyncGenerator<ExecutedWorkflowResponse<T>, void, unknown> {
   let data = await getWorkflowExecOutput<ExecutedWorkflowResponse<T>>(workflowId, abortController);
-  while (data.result.status === 'RUNNING') {
+  while (data?.result.status === 'RUNNING') {
+    // eslint-disable-next-line no-await-in-loop
+    await new Promise((resolve) => {
+      setTimeout(resolve, 800);
+    });
     yield data;
     // eslint-disable-next-line no-await-in-loop
     data = await getWorkflowExecOutput(workflowId, abortController);
   }
   // we need to do an additional yield for the last task status change
-  if (data.result.status === 'FAILED' || data.result.status === 'COMPLETED') {
+  if (data?.result.status === 'FAILED' || data?.result.status === 'COMPLETED') {
     yield data;
     if (onFinish) {
       onFinish(data.result.status === 'COMPLETED');
@@ -99,28 +106,26 @@ export type UseAsyncGeneratorParams = {
 };
 
 export function useAsyncGenerator<T>(props: UseAsyncGeneratorParams): ExecutedWorkflowPayload<T> | null {
-  const { current: controller } = useRef(new AbortController());
   const [execPayload, setExecPayload] = useState<ExecutedWorkflowPayload<T> | null>(null);
+  const { workflowId, onFinish } = props;
 
   useEffect(() => {
+    const controller = new AbortController();
     (async () => {
-      const { workflowId, onFinish } = props;
+      // we have to use async iterator here, so we turn off this rule
       if (workflowId != null) {
-        // we have to use async iterator here, so we turn off this rule
         // eslint-disable-next-line no-restricted-syntax
         for await (const data of asyncGenerator<T>({ workflowId, abortController: controller, onFinish })) {
           setExecPayload(data.result);
         }
       }
     })();
-  }, [props, controller]);
 
-  // we need to abort all fetch requests on unmount, otherwise we will get an error
-  useEffect(() => {
     return () => {
+      // we want to abort the request when user leaves view and/or changes the displayed workflow
       controller.abort();
     };
-  }, [controller]);
+  }, [workflowId, onFinish]);
 
   return execPayload;
 }
