+++ packages/frinx-gamma/src/pages/control-page/control-page-table.tsx
@@ -1,10 +1,7 @@
-import { Box, Button, Table, Tbody, Td, Th, Thead, Tr } from '@chakra-ui/react';
+import { Button, Table, Tbody, Td, Th, Thead, Tr } from '@chakra-ui/react';
 import React, { VoidFunctionComponent } from 'react';
 import { Link } from 'react-router-dom';
-import ControlPageTableButtons from './control-page-table-buttons';
-import WorkflowProgressBar from './workflow-progress-bar';
 
-type WorkflowState = { type: 'service' | 'bearer'; id: string };
 type CountState = {
   services: number;
   sites: number;
@@ -17,31 +14,15 @@ type TotalCountState = {
   deleted: CountState | null;
 };
 type Props = {
-  // onServicesSiteLinkClick: () => void;
-  // onSitesSiteLinkClick: () => void;
-  // onVpnBearerLinkClick: () => void;
-  workflowState: WorkflowState | null;
   countState: TotalCountState;
-  onServiceCommitBtnClick: () => void;
-  onBearerCommitBtnClick: () => void;
-  onWorkflowFinish: (isCompleted: boolean) => void;
 };
 
-const ControlPageTable: VoidFunctionComponent<Props> = ({
-  // onServicesSiteLinkClick,
-  // onSitesSiteLinkClick,
-  // onVpnBearerLinkClick,
-  workflowState,
-  countState,
-  onServiceCommitBtnClick,
-  onBearerCommitBtnClick,
-  onWorkflowFinish,
-}) => {
+const ControlPageTable: VoidFunctionComponent<Props> = ({ countState }) => {
   return (
     <Table background="white" size="lg">
       <Thead>
         <Tr>
-          <Th colSpan={2} />
+          <Th />
           <Th textAlign="right">Total</Th>
           <Th textAlign="right">Added</Th>
           <Th textAlign="right">Updated</Th>
@@ -55,13 +36,6 @@ const ControlPageTable: VoidFunctionComponent<Props> = ({
               Services
             </Button>
           </Td>
-          <Td minWidth={80}>
-            <Box minHeight={12}>
-              {workflowState?.type === 'service' && (
-                <WorkflowProgressBar onWorkflowFinish={onWorkflowFinish} workflowId={workflowState.id} />
-              )}
-            </Box>
-          </Td>
           <Td fontFamily="monospace" textAlign="right">
             {countState.total?.services ?? '-'}
           </Td>
@@ -74,25 +48,12 @@ const ControlPageTable: VoidFunctionComponent<Props> = ({
           <Td fontFamily="monospace" textAlign="right">
             {countState.deleted?.services ?? '-'}
           </Td>
-          <Td>
-            <ControlPageTableButtons
-              onCommitBtnClick={onServiceCommitBtnClick}
-              isCommitLoading={workflowState?.type === 'service'}
-            />
-          </Td>
         </Tr>
         <Tr>
           <Td>
             <Button variant="link" as={Link} to="../sites">
               Sites
             </Button>
-          </Td>
-          <Td minWidth={80}>
-            <Box minHeight={12}>
-              {workflowState?.type === 'service' && (
-                <WorkflowProgressBar onWorkflowFinish={onWorkflowFinish} workflowId={workflowState.id} />
-              )}
-            </Box>
           </Td>
           <Td fontFamily="monospace" textAlign="right">
             {countState.total?.sites ?? '-'}
@@ -106,25 +67,12 @@ const ControlPageTable: VoidFunctionComponent<Props> = ({
           <Td fontFamily="monospace" textAlign="right">
             {countState.deleted?.sites ?? '-'}
           </Td>
-          <Td>
-            <ControlPageTableButtons
-              onCommitBtnClick={onServiceCommitBtnClick}
-              isCommitLoading={workflowState?.type === 'service'}
-            />
-          </Td>
         </Tr>
         <Tr>
           <Td>
             <Button variant="link" as={Link} to="../vpn-bearers">
               VPN Bearers
             </Button>
-          </Td>
-          <Td minWidth={80}>
-            <Box minHeight={12}>
-              {workflowState?.type === 'bearer' && (
-                <WorkflowProgressBar onWorkflowFinish={onWorkflowFinish} workflowId={workflowState.id} />
-              )}
-            </Box>
           </Td>
           <Td fontFamily="monospace" textAlign="right">
             {countState.total?.bearers ?? '-'}
@@ -137,12 +85,6 @@ const ControlPageTable: VoidFunctionComponent<Props> = ({
           </Td>
           <Td fontFamily="monospace" textAlign="right">
             {countState.deleted?.bearers ?? '-'}
-          </Td>
-          <Td>
-            <ControlPageTableButtons
-              onCommitBtnClick={onBearerCommitBtnClick}
-              isCommitLoading={workflowState?.type === 'bearer'}
-            />
           </Td>
         </Tr>
       </Tbody>
