+++ packages/frinx-gamma/src/pages/control-page/control-page.tsx
@@ -1,24 +1,14 @@
-import { Box, Container, Flex, Heading, Icon, Button, Text, Tooltip, useDisclosure } from '@chakra-ui/react';
-import React, { useEffect, useState, VoidFunctionComponent } from 'react';
-import FeatherIcon from 'feather-icons-react';
-import uniflowCallbackUtils from '../../uniflow-callback-utils';
-import unistoreCallbackUtils from '../../unistore-callback-utils';
-import { getTransactionId, setTransactionId, removeTransactionId } from '../../helpers/transaction-id';
+import { Box, Container, Flex, Heading } from '@chakra-ui/react';
+import React, { VoidFunctionComponent } from 'react';
+import { CalcDiffPayload } from '../../components/commit-status-modal/commit-status-modal.helpers';
+import useCalcDiffContext from '../../providers/calcdiff-provider/user-calcdiff-context';
 import ControlPageTable from './control-page-table';
-import unwrap from '../../helpers/unwrap';
-import {
-  CalcDiffPayload,
-  ExecutedWorkflowPayload,
-  useAsyncGenerator,
-} from '../../components/commit-status-modal/commit-status-modal.helpers';
-import DiscardChangesModal from '../../components/discard-changes-modal/discard-changes';
 
 type CountState = {
   services: number;
   sites: number;
   bearers: number;
 };
-type WorkflowState = { type: 'service' | 'bearer'; id: string };
 type TotalCountState = {
   total: CountState | null;
   added: CountState | null;
@@ -32,12 +22,9 @@ const DEFAULT_UNCOMMITED_CHANGES: TotalCountState = {
   deleted: null,
 };
 
-function makeTotalCountState(
-  countState: TotalCountState,
-  payload: ExecutedWorkflowPayload<CalcDiffPayload> | null,
-): TotalCountState {
-  if (payload?.status === 'COMPLETED') {
-    const { changes } = payload.output;
+function makeTotalCountState(countState: TotalCountState, payload: CalcDiffPayload | null): TotalCountState {
+  if (payload != null) {
+    const { changes } = payload;
     return {
       total: countState.total,
       added: {
@@ -61,175 +48,25 @@ function makeTotalCountState(
 }
 
 const ControlPage: VoidFunctionComponent = () => {
-  const discardChangesDisclosure = useDisclosure();
-  const [countState, setCountState] = useState<TotalCountState>(DEFAULT_UNCOMMITED_CHANGES);
-  const [workflowState, setWorkflowState] = useState<WorkflowState | null>(null);
-  const [workflowId, setWorkflowId] = useState<string | null>(null);
-  const [isCommitedOrDiscarded, setIsCommitedOrDiscarded] = useState(false);
-  const onFinish = () => {
-    setWorkflowId(null);
-  };
-  const workflowPayload = useAsyncGenerator<CalcDiffPayload>({ workflowId, onFinish });
-
-  useEffect(() => {
-    (async () => {
-      const callbacks = unistoreCallbackUtils.getCallbacks;
-      const [serviceCount, siteCount, bearerCount] = await Promise.all([
-        callbacks.getVpnServiceCount(null),
-        callbacks.getVpnSiteCount(null),
-        callbacks.getVpnBearerCount(null),
-      ]);
-      setCountState((prev) => ({
-        ...prev,
-        total: {
-          services: serviceCount,
-          sites: siteCount,
-          bearers: bearerCount,
-        },
-      }));
-    })();
-  }, [isCommitedOrDiscarded]);
-
-  useEffect(() => {
-    const callbacks = uniflowCallbackUtils.getCallbacks;
-
-    callbacks
-      .executeWorkflow({
-        name: 'Process_calcdiff_ui',
-        version: 2,
-        input: {
-          // eslint-disable-next-line @typescript-eslint/naming-convention
-          unistore_node_name: 'service',
-          US_UI_TX: unwrap(getTransactionId()),
-        },
-      })
-      .then((data) => {
-        setWorkflowId(data.text);
-      });
-  }, []);
-
-  const handleServiceCommitBtnClick = () => {
-    const callbacks = uniflowCallbackUtils.getCallbacks;
-    callbacks
-      .executeWorkflow({
-        name: 'Render_all',
-        version: 1,
-        input: {
-          // eslint-disable-next-line @typescript-eslint/naming-convention
-          unistore_node_name: 'service',
-          action: 'dryrun',
-          US_UI_TX: unwrap(getTransactionId()),
-        },
-      })
-      .then((data) => {
-        setWorkflowState({
-          type: 'service',
-          id: data.text,
-        });
-        setIsCommitedOrDiscarded(true);
-      });
-  };
-
-  const handleBearerCommitBtnClick = () => {
-    const callbacks = uniflowCallbackUtils.getCallbacks;
-    callbacks
-      .executeWorkflow({
-        name: 'Render_bearer',
-        version: 2,
-        input: {
-          // eslint-disable-next-line @typescript-eslint/naming-convention
-          unistore_node_name: 'bearer',
-          action: 'dryrun',
-          US_UI_TX: unwrap(getTransactionId()),
-        },
-      })
-      .then((data) => {
-        setWorkflowState({
-          type: 'bearer',
-          id: data.text,
-        });
-        setIsCommitedOrDiscarded(true);
-      });
-  };
-
-  const handleDiscardConfirmBtnClick = async () => {
-    const callbacks = unistoreCallbackUtils.getCallbacks;
-    // TODO: close transaction endpoint response is empty 200, so our helper function fails to parse it
-    // we can handle it in our api-helpers.ts or find other solution instead of try/catch
-    try {
-      await callbacks.closeTransaction();
-    } catch (e) {
-      removeTransactionId();
-      const data = await callbacks.getTransactionCookie();
-      setTransactionId(data);
-      discardChangesDisclosure.onClose();
-      setIsCommitedOrDiscarded(true);
-    }
-  };
-
-  const handleDiscardBtnClick = () => {
-    discardChangesDisclosure.onOpen();
-  };
-
-  const handleWorkflowFinish = (isCompleted: boolean) => {
-    setWorkflowState(null);
-
-    if (!isCompleted) {
-      return;
-    }
-
-    // set new transaction id after successful commit
-    removeTransactionId();
-    const unistoreCallbacks = unistoreCallbackUtils.getCallbacks;
-    unistoreCallbacks.getTransactionCookie().then((data) => {
-      setTransactionId(data);
-    });
-  };
-
-  const uncommitedChanges = isCommitedOrDiscarded
-    ? DEFAULT_UNCOMMITED_CHANGES
-    : makeTotalCountState(countState, workflowPayload);
+  const { data } = useCalcDiffContext();
+  const countState = makeTotalCountState(DEFAULT_UNCOMMITED_CHANGES, data);
 
   return (
-    <>
-      <DiscardChangesModal
-        isOpen={discardChangesDisclosure.isOpen}
-        onClose={discardChangesDisclosure.onClose}
-        onConfirmBtnClick={handleDiscardConfirmBtnClick}
-        title="Discard changes"
-      >
-        Are you sure you want to discard your changes?
-      </DiscardChangesModal>
-      <Container maxWidth={1280} minHeight="60vh">
-        <Flex justify="space-between" align="center" marginBottom={6}>
-          <Heading as="h2" size="lg">
-            Control page
-          </Heading>
-          <Tooltip label="Discard changes">
-            <Button
-              aria-label="Discard changes"
-              leftIcon={<Icon as={FeatherIcon} icon="trash" />}
-              onClick={handleDiscardBtnClick}
-              colorScheme="blue"
-            >
-              <Text>Discard Changes</Text>
-            </Button>
-          </Tooltip>
-        </Flex>
-        <Box>
-          <ControlPageTable
-            countState={{
-              ...uncommitedChanges,
-              total: countState.total,
-            }}
-            workflowState={workflowState}
-            onServiceCommitBtnClick={handleServiceCommitBtnClick}
-            onBearerCommitBtnClick={handleBearerCommitBtnClick}
-            onWorkflowFinish={handleWorkflowFinish}
-          />
-        </Box>
-      </Container>
-    </>
+    <Container maxWidth={1280} minHeight="60vh">
+      <Flex justify="space-between" align="center" marginBottom={6}>
+        <Heading as="h2" size="lg">
+          Control page
+        </Heading>
+      </Flex>
+      <Box>
+        <ControlPageTable
+          countState={{
+            ...countState,
+            total: countState.total,
+          }}
+        />
+      </Box>
+    </Container>
   );
 };
