+++ packages/frinx-gamma/src/providers/calcdiff-provider/user-calcdiff-context.ts
@@ -1,0 +1,7 @@
+import { useContext } from 'react';
+import unwrap from '../../helpers/unwrap';
+import { CalcDiffContext, CalcDiffContextProps } from './calcdiff-provider';
+
+export default function useCalcDiffContext(): CalcDiffContextProps {
+  return unwrap(useContext(CalcDiffContext));
+}
